+++ include/NeoFOAM/dsl/operator.hpp
@@ -32,7 +32,7 @@ concept HasExplicitOperator = requires(T t) {
 };
 
 /* @class Operator
- * @brief A class to represent a operator in NeoFOAMs dsl
+ * @brief A class to represent an operator in NeoFOAMs dsl
  *
  * The design here is based on the type erasure design pattern
  * see https://www.youtube.com/watch?v=4eeESJQk-mw