+++ include/NeoFOAM/timeIntegration/forwardEuler.hpp
@@ -4,20 +4,24 @@
 
 #pragma once
 
+#include "NeoFOAM/core/database/fieldCollection.hpp"
+#include "NeoFOAM/core/database/oldTimeCollection.hpp"
 #include "NeoFOAM/fields/field.hpp"
 #include "NeoFOAM/timeIntegration/timeIntegration.hpp"
 
-namespace NeoFOAM::dsl
+namespace NeoFOAM::timeIntegration
 {
 
-template<typename SolutionType>
+template<typename SolutionFieldType>
 class ForwardEuler :
-    public TimeIntegratorBase<SolutionType>::template Register<ForwardEuler<SolutionType>>
+    public TimeIntegratorBase<SolutionFieldType>::template Register<ForwardEuler<SolutionFieldType>>
 {
 
 public:
 
-    using Base = TimeIntegratorBase<SolutionType>::template Register<ForwardEuler<SolutionType>>;
+    using Expression = NeoFOAM::dsl::Expression;
+    using Base =
+        TimeIntegratorBase<SolutionFieldType>::template Register<ForwardEuler<SolutionFieldType>>;
 
     ForwardEuler(const Dictionary& dict) : Base(dict) {}
 
@@ -27,21 +31,26 @@ public:
 
     static std::string schema() { return "none"; }
 
-    void solve(Expression& eqn, SolutionType& sol, scalar dt) override
+    void solve(
+        Expression& eqn, SolutionFieldType& solutionField, [[maybe_unused]] scalar t, scalar dt
+    ) override
     {
-        auto source = eqn.explicitOperation(sol.size());
+        auto source = eqn.explicitOperation(solutionField.size());
+        SolutionFieldType& oldSolutionField =
+            NeoFOAM::finiteVolume::cellCentred::oldTime(solutionField);
 
-        sol.internalField() -= source * dt;
-        sol.correctBoundaryConditions();
+        solutionField.internalField() = oldSolutionField.internalField() - source * dt;
+        solutionField.correctBoundaryConditions();
 
         // check if executor is GPU
         if (std::holds_alternative<NeoFOAM::GPUExecutor>(eqn.exec()))
         {
             Kokkos::fence();
         }
+        oldSolutionField.internalField() = solutionField.internalField();
     };
 
-    std::unique_ptr<TimeIntegratorBase<SolutionType>> clone() const override
+    std::unique_ptr<TimeIntegratorBase<SolutionFieldType>> clone() const override
     {
         return std::make_unique<ForwardEuler>(*this);
     }