+++ test/finiteVolume/cellCentred/boundary/surface/surfFixedValue.cpp
@@ -21,7 +21,7 @@ TEST_CASE("fixedValue")
         NeoFOAM::Executor(NeoFOAM::GPUExecutor {})
     );
 
-    std::string execName = std::visit([](auto e) { return e.print(); }, exec);
+    std::string execName = std::visit([](auto e) { return e.name(); }, exec);
 
     SECTION("TestDerivedClass" + execName)
     {