+++ test/finiteVolume/cellCentred/interpolation/linear.cpp
@@ -25,7 +25,7 @@ TEST_CASE("linear")
         NeoFOAM::Executor(NeoFOAM::GPUExecutor {})
     );
 
-    std::string execName = std::visit([](auto e) { return e.print(); }, exec);
+    std::string execName = std::visit([](auto e) { return e.name(); }, exec);
     auto mesh = NeoFOAM::createSingleCellMesh(exec);
     NeoFOAM::Input input = NeoFOAM::TokenList({std::string("linear")});
     auto linear = SurfaceInterpolation(exec, mesh, input);